extract the single element serialization idiom into a serializer extension
